Using better way of specifying fixture state (concordion:state=ExpectedToFail).
