+++ src/main/java/org/concordion/api/FixtureState.java
@@ -1,0 +1,10 @@
+package org.concordion.api;
+
+/**
+ * Created by tim on 7/07/15.
+ */
+public enum FixtureState {
+    Unimplemented,
+    ExpectedToPass,
+    ExpectedToFail;
+}
