+++ src/main/java/org/concordion/internal/command/ExampleCommand.java
@@ -2,9 +2,8 @@ package org.concordion.internal.command;
 
 import org.concordion.api.*;
 import org.concordion.internal.FailFastException;
-import org.concordion.internal.FixtureState;
+import org.concordion.internal.InternalFixtureState;
 import org.concordion.internal.SpecificationDescriber;
-import org.concordion.internal.listener.SpecificationExporter;
 
 import java.util.Arrays;
 import java.util.List;
@@ -35,15 +34,25 @@ public class ExampleCommand extends AbstractCommand {
             aName.addAttribute("id", "#" + node.getExpression()); // html5 version
             node.getElement().prependChild(aName);
 
-            // let's be really nice and add the fixture state text into the element itself.
-            FixtureState state = FixtureState.getFixtureState(
-                    null,
-                    node.getExpression());
+            String params = node.getParameter("state");
+            if (params != null) {
+                FixtureState fixtureState = FixtureState.valueOf(params);
+                resultRecorder.setFixtureState(fixtureState);
+                // let's be really nice and add the fixture state text into the element itself.
+                InternalFixtureState internalFixtureState = InternalFixtureState.getFixtureState(null, fixtureState);
 
-            String note = state.printNoteToString();
-            Element fixtureNode = new Element("p");
-            fixtureNode.appendText(note);
-            node.getElement().prependChild(fixtureNode);
+                String note;
+                if (internalFixtureState != null) {
+                    note = internalFixtureState.printNoteToString();
+                } else if (internalFixtureState == null) {
+                    note = "Invalid state expression " + params;
+                } else {
+                    note = "";
+                }
+                Element fixtureNode = new Element("p");
+                fixtureNode.appendText(note);
+                node.getElement().prependChild(fixtureNode);
+            }
 
         } catch (FailFastException e) {
             // Ignore - it'll be re-thrown later if necessary.
